+++ lib/startback/web.rb
@@ -1,4 +1,5 @@
 require_relative 'web/shield'
 require_relative 'web/catch_all'
 require_relative 'web/health_check'
+require_relative 'web/auto_caching'
 require_relative 'web/api'